+++ templates/word-daily-level/index.html
@@ -22,11 +22,11 @@
   </div>
 
   <div class="screen" id="playingScreen">
-    <h2 id="prompt"></h2>
+    <h2 id="prompt">Try to guess today's N letter word</h2>
     <h2 id="answer"></h2>
     <h4 id="hint"></h4>
 
-    <word-game-grid></word-game-grid>
+    <word-game-grid max-attempts="6"></word-game-grid>
     <word-game-keyboard></word-game-keyboard>
   </div>
