--- templates/word-daily-level/src/consts.js
@@ -1,1 +1,0 @@
-export const ALLOWED_ATTEMPTS = 6;
